+++ CHANGES.md
@@ -4,6 +4,7 @@ This files lists the changes during the lifetime of this project.
 
 ## unreleased
 
+- 261: (new env var) move db admin from `/djadmin/db/` to `/staff/`, use SSO login instead of superuser login, access controlled by `STAFF_EMAILS`
 - 262: make sidepanels the same between 2 pages
 - 262: (migration) change status from assignment to service: (CONCEPT, OPEN, GESLOTEN)
 - 251: fix that placement table shows org labels instead of org names